requirements.txt
fastapi==0.68.0
uvicorn==0.12.2
python-dotenv~=0.21.0
pydantic~=1.10.2
SQLAlchemy~=1.4.42
mysqlclient
httpx
requests